Replace client-side Discogs markup parser with server-provided tokens
The Backend-Service now returns bioTokens (pre-parsed structured tokens from LML) alongside the raw bio text. Consume these tokens directly in the renderer instead of parsing markup client-side.

Delete discogsMarkup.ts (parser) and discogsMarkup.test.ts (419 lines of tests) since the server handles parsing and entity resolution. The renderer falls back to plain bio text when tokens are unavailable.
