+++ lib/features/metadata/types.ts
@@ -18,9 +18,21 @@ export interface AlbumMetadata {
   artistWikipediaUrl?: string;
 }
 
+export type ResolvedToken =
+  | { type: "plainText"; text: string }
+  | { type: "artistLink"; name: string; display_name: string; url: string }
+  | { type: "labelName"; name: string }
+  | { type: "releaseLink"; title: string; url: string }
+  | { type: "masterLink"; title: string; url: string }
+  | { type: "bold"; content: string }
+  | { type: "italic"; content: string }
+  | { type: "underline"; content: string }
+  | { type: "urlLink"; href: string | null; content: string };
+
 export interface ArtistMetadata {
   discogsArtistId: number;
   bio: string | null;
+  bioTokens: ResolvedToken[] | null;
   wikipediaUrl: string | null;
   imageUrl: string | null;
 }
